infrastructure/requirements.txt
aws-cdk-lib>=2.110.0
aws-cdk.aws-glue-alpha==2.110.0a0
constructs>=10.0.0,<11.0.0
